+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Add Bisu Platform as a translation source
+
 ## [3.0.2](https://github.com/hole19/bisu/releases/tag/v3.0.2)(2025-11-26)
 
 - Escape percentage character properly on Android